completing edit plant details logic

+++ lib/features/home/presentation/views/widgets/plantDetails/plantDetailItem.dart
@@ -29,7 +29,7 @@ class PlantDetailItem extends StatelessWidget {
         ),
         const SizedBox(height: 10,),
         FancyProgressBar(
-          progress: percentage,
+          progress: percentage.toDouble(),
         ),
       ],
     );
